chore(contest rules): move ExamTable.tsx to rule folder
