+++ src/components/contest/rules/ExamTable.tsx
